Move ios specific modules behind shared cfg
