+++ mullvad-ios/src/ios/api_client/completion.rs
