+++ mullvad-ios/src/ios/api_client/device.rs
