+++ mullvad-ios/src/ios/api_client/helpers.rs
