+++ mullvad-ios/src/ios/logging.rs
